dev/u2a/auth/requirements.txt
aiocache==0.11.1
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
circuitbreaker==1.3.2
click==8.1.2
Deprecated==1.2.13
distlib==0.3.4
elasticsearch==7.9.1
fastapi==0.74.1
filelock==3.6.0
flake8==4.0.1
flasgger==0.9.5
Flask==2.1.1
Flask-JWT-Extended==4.4.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
gevent==21.12.0
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
identify==2.5.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.5.1
lasier==0.2.3
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mccabe==0.6.1
mistune==2.0.2
multidict==6.0.2
nodeenv==1.6.0
oauthlib==3.2.0
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
orjson==3.6.5
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.8
pyrate-limiter==2.8.1
pyrsistent==0.18.1
pytest==7.1.2
pytest-asyncio==0.12.0
pytest-repeat==0.9.1
python-dotenv==0.19.2
PyYAML==6.0
redis==4.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
thrift==0.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.15.3
virtualenv==20.14.1
Werkzeug==2.1.1
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
